refactor: change files again and declaration after router is created
