+++ src/App.tsx
@@ -4,8 +4,8 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import { useAuth, type AuthContext } from "react-oidc-context";
 
-import { OidcAuthProvider } from "./routes/OidcAuthProvider";
-import { router } from "./routes/router";
+import { OidcAuthProvider } from "./components/OidcAuthProvider";
+import { router } from "./router.config";
 import "./i18n.config";
 
 // Create a client for react-query
