+++ src/pages/AgreementsSearch/AgreementsSearchPage.tsx
@@ -3,8 +3,8 @@ import { Link, useSearch } from "@tanstack/react-router";
 import { createColumnHelper } from "@tanstack/react-table";
 import { useTranslation } from "react-i18next";
 
-import { searchAgreementsRoute } from "../../routes/routing/agreements/searchAgreements";
-import Protector from "../../routes/Protector";
+import { searchAgreementsRoute } from "../../routes/agreements/searchAgreements";
+import Protector from "../../components/Protector";
 import ModuleTable, {
   type ColumnVisibilityGraph,
 } from "../../components/PrimaryModule/ModuleTable";
