+++ src/pages/CustomerSearch/CustomerSearchPage.tsx
@@ -3,8 +3,8 @@ import { Link, useSearch } from "@tanstack/react-router";
 import { createColumnHelper } from "@tanstack/react-table";
 import { useTranslation } from "react-i18next";
 
-import { searchCustomersRoute } from "../../routes/routing/customers/searchCustomers";
-import Protector from "../../routes/Protector";
+import { searchCustomersRoute } from "../../routes/customers/searchCustomers";
+import Protector from "../../components/Protector";
 import ModuleTable, {
   type ColumnVisibilityGraph,
 } from "../../components/PrimaryModule/ModuleTable";
