+++ src/pages/VehicleView/VehicleViewPage.tsx
@@ -6,7 +6,8 @@ import {
   useSearch,
 } from "@tanstack/react-router";
 
-import Protector from "../../routes/Protector";
+import { viewVehicleRoute } from "../../routes/vehicles/viewVehicle";
+import Protector from "../../components/Protector";
 import {
   ChevronLeftOutline,
   HamburgerMenuOutline,
@@ -19,7 +20,6 @@ import {
 import ScrollToTop from "../../components/ScrollToTop";
 
 import { getStartingIndexFromTabName } from "../../utils/moduleTabs";
-import { viewVehicleRoute } from "../../routes/routing/vehicles/viewVehicle";
 
 const SummaryTab = lazy(
   () => import("../../components/Vehicle/VehicleSummaryTab")
