+++ src/routes/customers/index.ts
