+++ src/routes/customers/searchCustomers.ts
