+++ src/routes/customers/viewCustomer.ts
