+++ src/routes/vehicles/viewVehicle.ts
